requirements.txt
arxiv==2.4.0
google-generativeai>=0.8.0
requests==2.31.0
PyYAML==6.0.1
pytest==8.0.0
pytest-mock==3.12.0
